+++ src/cli/config.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use clap::Parser;
 use colored::Colorize;
 use open;
-use std::fs::File;
+use std::fs::{self, File};
 
 use crate::{argon_info, config::Config as GlobalConfig, exit, ext::PathExt, logger, util};
 
@@ -20,12 +20,32 @@ pub struct Config {
 	/// List all available settings
 	#[arg(short, long)]
 	list: bool,
+
+	/// Restore all settings to default values
+	#[arg(short, long)]
+	default: bool,
 }
 
 impl Config {
 	pub fn main(self) -> Result<()> {
 		if self.list {
 			argon_info!("List of all available config options:\n\n{}", GlobalConfig::list());
+
+			return Ok(());
+		}
+
+		if self.default {
+			let config_path = util::get_argon_dir()?.join("config.toml");
+
+			if config_path.exists() {
+				if GlobalConfig::new().move_to_bin {
+					trash::delete(config_path)?;
+				} else {
+					fs::remove_file(config_path)?;
+				}
+			}
+
+			argon_info!("Restored all settings to default values");
 
 			return Ok(());
 		}